risk_intelligence/requirements.txt
# ── Core Server ─────────────────────────────────────
fastapi==0.133.1
uvicorn==0.41.0
starlette==0.52.1
python-dotenv==1.2.1
python-multipart==0.0.22
jinja2==3.1.6
click==8.1.8
h11==0.16.0

# ── Data Validation ─────────────────────────────────
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2

# ── LLM & Agents ───────────────────────────────────
langchain==1.2.10
langchain-core==1.2.16
langchain-community==0.4.1
langchain-groq==1.1.2
langchain-text-splitters==1.1.1
langchain-google-genai==4.2.1
langchain-classic==1.0.1
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.7
groq==0.37.1
openai==2.16.0
google-genai==1.65.0

# ── RAG / Vector Store ─────────────────────────────
chromadb>=1.0.0
faiss-cpu==1.13.2

# ── HTTP / Networking ──────────────────────────────
httpx==0.28.1
httpcore==1.0.9
httpx-sse==0.4.3
aiohttp==3.13.3
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
anyio==4.12.1
sniffio==1.3.1
requests==2.32.5
requests-toolbelt==1.0.0
urllib3==2.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
websockets==16.0
multidict==6.7.1
frozenlist==1.8.0
yarl==1.22.0
propcache==0.4.1

# ── Data Processing ────────────────────────────────
pandas==2.3.3
numpy==2.4.2
openpyxl>=3.1.0
pyarrow==23.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.3
six==1.17.0

# ── PDF Processing ─────────────────────────────────
pypdf==6.7.3
pillow==12.1.1

# ── JSON / Serialization ──────────────────────────
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0
orjson==3.11.7
ormsgpack==1.12.2
jiter==0.12.0
dataclasses-json==0.6.7
marshmallow==3.26.2
mypy_extensions==1.1.0

# ── Database ───────────────────────────────────────
SQLAlchemy==2.0.45

# ── Web Scraping / Parsing ─────────────────────────
beautifulsoup4==4.14.3
soupsieve==2.8.3
wikipedia==1.4.0
filetype==1.2.0

# ── Auth / Crypto ──────────────────────────────────
google-auth==2.48.0
cryptography==46.0.5
cffi==2.0.0
pycparser==3.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
rsa==4.9.1
cachetools==6.2.6

# ── Flask (if needed for secondary apps) ───────────
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
itsdangerous==2.2.0
blinker==1.9.0

# ── Streamlit (if needed for dashboards) ───────────
streamlit==1.54.0
altair==6.0.0
pydeck==0.9.1
toml==0.10.2
tornado==6.5.4
narwhals==2.17.0
Markdown==3.10.2

# ── Git / Env / Utils ─────────────────────────────
GitPython==3.1.46
gitdb==4.0.12
smmap==5.0.2
virtualenv==20.31.2
distlib==0.3.9
filelock==3.18.0
platformdirs==4.3.8
distro==1.9.0
packaging==26.0
tqdm==4.67.2
tenacity==9.1.4
PyYAML==6.0.3
MarkupSafe==3.0.2
attrs==25.4.0
annotated-doc==0.0.4
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
community==1.0.0b1
protobuf==6.33.5
